compiled main with -Wconversion

+++ src/LLC/types/uint1024.h
@@ -341,10 +341,10 @@ public:
         for(int index = 0; index < WIDTH; ++index)
         {
             std::vector<uint8_t> BYTES(4, 0);
-            BYTES[0] = (pn[index] >> 24);
-            BYTES[1] = (pn[index] >> 16);
-            BYTES[2] = (pn[index] >> 8);
-            BYTES[3] =  pn[index];
+            BYTES[0] = static_cast<uint8_t>(pn[index] >> 24);
+            BYTES[1] = static_cast<uint8_t>(pn[index] >> 16);
+            BYTES[2] = static_cast<uint8_t>(pn[index] >> 8);
+            BYTES[3] = static_cast<uint8_t>(pn[index]);
 
             DATA.insert(DATA.end(), BYTES.begin(), BYTES.end());
         }
